+++ compose-guard/gradle-plugin/src/main/kotlin/com/joetr/compose/guard/ReportGenPlugin.kt
@@ -176,6 +176,7 @@ public class ReportGenPlugin : Plugin<Project> {
                 hasKotlinMainSourceSet.set(
                     project.hasNonEmptyKotlinSourceSets(),
                 )
+                reportAllOnMissingBaseline.set(checkExtension.reportAllOnMissingBaseline)
 
                 kotlinSourceSets.set(project.getKotlinSources())
                 taskNameProperty.set(taskName)